# contracts/internal/agent-openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "AI 专家平台 · 内部 Agent API",
    "description": "仅供 backend 内网调用。永不对公网开放。",
    "version": "0.0.1"
  },
  "paths": {
    "/internal/health": {
      "get": {
        "tags": [
          "system"
        ],
        "summary": "存活探针（不校验 token，供编排系统探活）",
        "operationId": "health_internal_health_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HealthResponse"
                }
              }
            }
          }
        }
      }
    },
    "/internal/extract": {
      "post": {
        "tags": [
          "ingest"
        ],
        "summary": "把链接或文件提取成纯文本（同步）",
        "description": "同步而非入队：抓不到链接、解析不了文件应该在博主点上传的那一刻就告诉他，让他改用粘贴正文。失败时返回 422 且 detail 是可直接展示给用户的中文原因。",
        "operationId": "extract_content_internal_extract_post",
        "parameters": [
          {
            "name": "x-internal-token",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Internal-Token"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExtractRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExtractResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/internal/chat": {
      "post": {
        "tags": [
          "chat"
        ],
        "summary": "召回 + 双闸门 + 流式生成（SSE）",
        "description": "以 text/event-stream 返回，协议见 contracts/README.md。\n\n入口闸门：rerank 分数低于阈值的召回结果全部丢弃；一条不剩时直接返回「这个他没有讲过」，不调用生成模型。\n\n出口闸门：流式下做不到边流边审，所以先流给用户、同时缓冲全文，流结束后校验一次禁区，命中则追加免责说明。",
        "operationId": "chat_internal_chat_post",
        "parameters": [
          {
            "name": "x-internal-token",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Internal-Token"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChatRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              },
              "text/event-stream": {}
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/internal/extract-model": {
      "post": {
        "tags": [
          "build"
        ],
        "summary": "只重新提炼七维，不重新向量化",
        "description": "复用已有的知识切片，只跑提炼。博主会反复重新生成直到满意，每次都重跑 embedding 是真金白银。",
        "operationId": "extract_model_internal_extract_model_post",
        "parameters": [
          {
            "name": "x-internal-token",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Internal-Token"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExtractModelRequest"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BuildAccepted"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/internal/build": {
      "post": {
        "tags": [
          "build"
        ],
        "summary": "入队构建任务，立即返回 job_id",
        "description": "进度写入 build_jobs 表，由 backend 读给前端轮询。任务失败时必须把 status 置为 failed 并写入可读的 error。",
        "operationId": "build_internal_build_post",
        "parameters": [
          {
            "name": "x-internal-token",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Internal-Token"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BuildRequest"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BuildAccepted"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/internal/readyz": {
      "get": {
        "tags": [
          "system"
        ],
        "summary": "就绪探针（检查数据库可达）",
        "operationId": "readyz_internal_readyz_get",
        "parameters": [
          {
            "name": "x-internal-token",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Internal-Token"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReadyResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "BuildAccepted": {
        "properties": {
          "job_id": {
            "type": "string",
            "format": "uuid",
            "title": "Job Id"
          }
        },
        "type": "object",
        "required": [
          "job_id"
        ],
        "title": "BuildAccepted"
      },
      "BuildRequest": {
        "properties": {
          "expert_id": {
            "type": "string",
            "format": "uuid",
            "title": "Expert Id"
          },
          "tenant_id": {
            "type": "string",
            "format": "uuid",
            "title": "Tenant Id"
          },
          "material_ids": {
            "items": {
              "type": "string",
              "format": "uuid"
            },
            "type": "array",
            "title": "Material Ids"
          }
        },
        "type": "object",
        "required": [
          "expert_id",
          "tenant_id"
        ],
        "title": "BuildRequest",
        "description": "由 backend 调用。tenant_id 虽然来自可信内网，agent 仍会用 experts 表核对\n（见 db/session.py 的 verified_tenant_conn）—— 可信不等于不会有 bug。"
      },
      "ChatRequest": {
        "properties": {
          "expert_id": {
            "type": "string",
            "format": "uuid",
            "title": "Expert Id"
          },
          "tenant_id": {
            "type": "string",
            "format": "uuid",
            "title": "Tenant Id"
          },
          "question": {
            "type": "string",
            "title": "Question"
          }
        },
        "type": "object",
        "required": [
          "expert_id",
          "tenant_id",
          "question"
        ],
        "title": "ChatRequest",
        "description": "tenant_id 由 backend 从分享短链反查得出，绝不来自客户端。\nagent 侧仍会用 experts 表再核对一次（纵深防御）。"
      },
      "ExtractModelRequest": {
        "properties": {
          "expert_id": {
            "type": "string",
            "format": "uuid",
            "title": "Expert Id"
          },
          "tenant_id": {
            "type": "string",
            "format": "uuid",
            "title": "Tenant Id"
          }
        },
        "type": "object",
        "required": [
          "expert_id",
          "tenant_id"
        ],
        "title": "ExtractModelRequest"
      },
      "ExtractRequest": {
        "properties": {
          "source_type": {
            "type": "string",
            "enum": [
              "file",
              "url"
            ],
            "title": "Source Type"
          },
          "url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Url"
          },
          "filename": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Filename"
          },
          "content_base64": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Content Base64"
          }
        },
        "type": "object",
        "required": [
          "source_type"
        ],
        "title": "ExtractRequest",
        "description": "paste 不走这里 —— 已经是文本了，backend 直接用，省一次网络往返。"
      },
      "ExtractResponse": {
        "properties": {
          "title": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Title"
          },
          "text": {
            "type": "string",
            "title": "Text"
          },
          "char_count": {
            "type": "integer",
            "title": "Char Count"
          }
        },
        "type": "object",
        "required": [
          "title",
          "text",
          "char_count"
        ],
        "title": "ExtractResponse"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "HealthResponse": {
        "properties": {
          "service": {
            "type": "string",
            "const": "agent",
            "title": "Service",
            "default": "agent"
          },
          "status": {
            "type": "string",
            "const": "ok",
            "title": "Status",
            "default": "ok"
          }
        },
        "type": "object",
        "title": "HealthResponse"
      },
      "ReadyResponse": {
        "properties": {
          "database": {
            "type": "string",
            "enum": [
              "ok",
              "down"
            ],
            "title": "Database"
          }
        },
        "type": "object",
        "required": [
          "database"
        ],
        "title": "ReadyResponse"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}